tostr.llm move to submodule under .semantic
