+++ src/tostr/semantic/llm/base.py
